Refactor components into individual folders
