+++ src/components/LanguageSelector/index.tsx
@@ -1,4 +1,4 @@
-import { useLanguage } from '../contexts/LanguageContext';
+import { useLanguage } from '../../contexts/LanguageContext';
 
 export default function LanguageSelector() {
   const { language, setLanguage } = useLanguage();
