+++ src/components/PDFUpload/index.tsx
@@ -1,6 +1,6 @@
 import { useState, useCallback } from 'react';
 import { useDropzone } from 'react-dropzone';
-import { useLanguage } from '../contexts/LanguageContext';
+import { useLanguage } from '../../contexts/LanguageContext';
 
 interface PDFUploadProps {
   onFileSelect: (file: File) => void;
