+++ src/components/layout/Header/index.tsx
@@ -1,4 +1,4 @@
-import { useLanguage } from '../../contexts/LanguageContext';
+import { useLanguage } from '../../../contexts/LanguageContext';
 
 export default function Header() {
   const { t } = useLanguage();
